refactor: standardize masthead filename convention
Rename toplevel-page__card-masthead.ts to {route}-card-masthead.ts:
- dataset/toplevel-page__card-masthead.ts → dataset-card-masthead.ts
- training/toplevel-page__card-masthead.ts → training-card-masthead.ts
- news-events/toplevel-page__card-masthead.ts → news-events-card-masthead.ts

Aligns with home/home-card-masthead.tsx pattern.
