+++ app/site-config/news-events/news-events-card-masthead.ts
