+++ app/site-config/training/training-card-masthead.ts
